Disabled demo mode, minor corrections
