+++ helpers.hpp
@@ -1,0 +1,8 @@
+#ifndef HELPERS_H
+#define HELPERS_H
+
+#include <string>
+
+std::string UTF8toISO8859_1(const char *in);
+
+#endif //HELPERS_H
