+++ printer.cpp
@@ -1,6 +1,7 @@
 #include "printer.hpp"
 #include "globals.hpp"
 #include "inputfield.hpp"
+#include "helpers.hpp"
 #include "messagesystem.h"
 #include <filesystem>
 #include <fstream>
@@ -119,36 +120,6 @@ bool PrinterRaw::printJPEG(const jpeg &jpeg_object) {
   return send_raw(output);
 }
 
-std::string UTF8toISO8859_1(const char *in) {
-  std::string out;
-  if (in == NULL)
-    return out;
-
-  unsigned int codepoint;
-  while (*in != 0) {
-    unsigned char ch = static_cast<unsigned char>(*in);
-    if (ch <= 0x7f)
-      codepoint = ch;
-    else if (ch <= 0xbf)
-      codepoint = (codepoint << 6) | (ch & 0x3f);
-    else if (ch <= 0xdf)
-      codepoint = ch & 0x1f;
-    else if (ch <= 0xef)
-      codepoint = ch & 0x0f;
-    else
-      codepoint = ch & 0x07;
-    ++in;
-    if (((*in & 0xc0) != 0x80) && (codepoint <= 0x10ffff)) {
-      if (codepoint <= 255) {
-        out.append(1, static_cast<char>(codepoint));
-      } else {
-        // do whatever you want for out-of-bounds characters
-      }
-    }
-  }
-  return out;
-}
-
 bool PrinterRaw::printLabel(label_info info) {
   std::stringstream output;
   output << "N\n";
@@ -179,16 +150,17 @@ bool PrinterRaw::printLabel(label_info info) {
   if (tokens.size() > 0)
     output << "A2," << linenumber * 20 + lineoffset << ",0,2,1,1,N,\"";
   for (int i = 0; i < tokens.size(); i++) {
-    if (linenumber > 2) linewidth = 22;
+    if (linenumber > 2)
+      linewidth = 22;
     int wordwidth =
         (tokens[i].size() <= linewidth - 1 ? tokens[i].size() : linewidth - 1);
     if (linepos + (wordwidth + 1) > linewidth) {
       linenumber++;
-      if (linenumber > 2)
-      {
+      if (linenumber > 2) {
         linewidth = 22;
-        wordwidth = (tokens[i].size() <= linewidth - 1 ? tokens[i].size() : linewidth - 1);
-      } 
+        wordwidth = (tokens[i].size() <= linewidth - 1 ? tokens[i].size()
+                                                       : linewidth - 1);
+      }
       output << "\"\n";
       output << "A2," << linenumber * 20 + lineoffset << ",0,2,1,1,N,\"";
       linepos = wordwidth;
@@ -214,10 +186,9 @@ bool PrinterRaw::printLabel(label_info info) {
     output << "B8,105,0,1,2,3,40,N,\"" + info.barcode + "\"\n";
   output << "P1\n";
 
-  //std::cout << "Output Buffer: \n" << output.str();
+  // std::cout << "Output Buffer: \n" << output.str();
   return send_raw(output.str());
-  //return true;
-
+  // return true;
 };
 
 bool PrinterRaw::openCashDrawer() {
@@ -239,28 +210,33 @@ PrinterRaw::PrinterRaw() {
 
 #ifdef __linux__
 
-PrinterLinuxUSBRAW::PrinterLinuxUSBRAW() : PrinterRaw() {
+PrinterLinuxUSBRAW::PrinterLinuxUSBRAW(bool label) : PrinterRaw() {
   name = "Linux USB Printer";
   addField(new string_field("device", "Device", "/dev/usb/lp0"), 0);
-  std::vector<std::string> protocol_type;
-  protocol_type.emplace(protocol_type.end(), std::string("escpos"));
-  protocol_type.emplace(protocol_type.end(), std::string("star"));
+  if (label)
+    GlobalState::registerLabelPrinter(name, this);
+  else {
+    std::vector<std::string> protocol_type;
+    protocol_type.emplace(protocol_type.end(), std::string("escpos"));
+    protocol_type.emplace(protocol_type.end(), std::string("star"));
 
-  addField(new string_combo_list_field("protocol_type", "Protocol",
-                                       std::move(protocol_type), 0),
-           1);
-  addField(new integer_range_field("gamma", "Gamma", 1000, 200, 1000, "slider"),
-           2);
-  addField(new integer_range_field("max_width", "Max Width", 576, 384, 576,
-                                   "slider"),
-           3);
-  addField(new integer_range_field("lines_to_feed", "Feed Lines", 0, 0, 20,
-                                   "spinbox"),
-           4);
-  addField(new boolean_field("paper_cut", "Cut Paper", false), 5);
-  addField(new boolean_field("cash_drawer", "Enable Cash Drawer", false), 6);
+    addField(new string_combo_list_field("protocol_type", "Protocol",
+                                         std::move(protocol_type), 0),
+             1);
+    addField(
+        new integer_range_field("gamma", "Gamma", 1000, 200, 1000, "slider"),
+        2);
+    addField(new integer_range_field("max_width", "Max Width", 576, 384, 576,
+                                     "slider"),
+             3);
+    addField(new integer_range_field("lines_to_feed", "Feed Lines", 0, 0, 20,
+                                     "spinbox"),
+             4);
+    addField(new boolean_field("paper_cut", "Cut Paper", false), 5);
+    addField(new boolean_field("cash_drawer", "Enable Cash Drawer", false), 6);
 
-  GlobalState::registerPrinter(name, this);
+    GlobalState::registerPrinter(name, this);
+  }
 }
 
 device_status PrinterLinuxUSBRAW::updateAndGetStatus() {
@@ -285,126 +261,6 @@ bool PrinterLinuxUSBRAW::send_raw(const std::string &buffer) {
   return true;
 }
 
-LabelPrinterLinuxUSBRAW::LabelPrinterLinuxUSBRAW() : PrinterRaw() {
-  name = "Linux USB Label Printer";
-  addField(new string_field("device", "Device", "/dev/usb/lp0"), 0);
-  GlobalState::registerLabelPrinter(name, this);
-}
-
-device_status LabelPrinterLinuxUSBRAW::updateAndGetStatus() {
-  return std::filesystem::exists(device()) ? CONNECTED : DISCONNECTED;
-}
-
-bool LabelPrinterLinuxUSBRAW::send_raw(const std::string &buffer) {
-  std::fstream file;
-  try {
-    file.exceptions(std::ofstream::badbit | std::ofstream::failbit);
-    file.open(fields.at("device")->get_string(), std::ios::in | std::ios::out);
-    file << buffer;
-  } catch (const std::ofstream::failure &e) {
-    std::cout << "Failure to open or write to Linux USB Raw Printer! Check "
-                 "permissions and address...\n";
-    if (file.is_open())
-      file.close();
-    return false;
-  }
-  if (file.is_open())
-    file.close();
-  return true;
-}
-
-std::string UTF8toISO8859_1(const char *in) {
-  std::string out;
-  if (in == NULL)
-    return out;
-
-  unsigned int codepoint;
-  while (*in != 0) {
-    unsigned char ch = static_cast<unsigned char>(*in);
-    if (ch <= 0x7f)
-      codepoint = ch;
-    else if (ch <= 0xbf)
-      codepoint = (codepoint << 6) | (ch & 0x3f);
-    else if (ch <= 0xdf)
-      codepoint = ch & 0x1f;
-    else if (ch <= 0xef)
-      codepoint = ch & 0x0f;
-    else
-      codepoint = ch & 0x07;
-    ++in;
-    if (((*in & 0xc0) != 0x80) && (codepoint <= 0x10ffff)) {
-      if (codepoint <= 255) {
-        out.append(1, static_cast<char>(codepoint));
-      } else {
-        // do whatever you want for out-of-bounds characters
-      }
-    }
-  }
-  return out;
-}
-
-bool LabelPrinterLinuxUSBRAW::printLabel(label_info info) {
-  std::stringstream output;
-  output << "N\n";
-
-  std::string pname = UTF8toISO8859_1(info.pname.c_str());
-
-  // Vector of string to save tokens
-  std::vector<std::string> tokens;
-
-  // stringstream class check1
-  std::stringstream check1(pname);
-  std::string intermediate;
-
-  // Tokenizing w.r.t. space ' '
-  while (getline(check1, intermediate, ' ')) {
-    tokens.push_back(intermediate);
-  }
-
-  // Printing the token vector
-  int linepos = 0;
-  int linenumber = 0;
-  int lineoffset = 3;
-
-  int linewidth = info.printprice ? 12 : 22;
-
-  if (tokens.size() > 0)
-    output << "A8," << linenumber * 20 + lineoffset << ",0,2,1,1,N,\"";
-  for (int i = 0; i < tokens.size(); i++) {
-    if (linenumber > 2) linewidth = 22;
-    int wordwidth =
-        (tokens[i].size() <= linewidth - 1 ? tokens[i].size() : linewidth - 1);
-    if (linepos + (wordwidth + 1) > linewidth) {
-      linenumber++;
-      output << "\"\n";
-      output << "A8," << linenumber * 20 + lineoffset << ",0,2,1,1,N,\"";
-      linepos = wordwidth;
-    } else
-      linepos += wordwidth;
-    for (int j = 0; j < wordwidth; j++) {
-      if (tokens[i][j] == '\"') {
-        output << "\\\"";
-      } else
-        output << tokens[i][j];
-    }
-    output << " ";
-  }
-  output << "\"\n";
-
-  if (info.printprice) {
-    output.precision(2);
-    output << "A160," << 20 << ",0,2,1,2,N,\"$ " << std::fixed << info.unitprice
-           << "\"\n";
-  }
-
-  if (info.barcode != "")
-    output << "B8,105,0,1,2,3,40,N,\"" + info.barcode + "\"\n";
-  output << "P1\n";
-
-  std::cout << "Output Buffer: \n" << output.str();
-  return send_raw(output.str());
-  //return true;
-}
 
 PrinterThermalLinuxTCPIP::PrinterThermalLinuxTCPIP() {
   name = "Linux TCP/IP Printer";
